production_app1/backend/requirements_real.txt
flask==2.2.3
flask-cors==3.0.10
requests==2.28.2
PySocks==1.7.1